+++ src/main/kotlin/projects/huegene/huegeneFlow.kt
@@ -127,7 +127,7 @@ fun huegeneFlow() = Sketch {
 
                         neighborhood.forEach { offset ->
                             val checkVector = chosenPoint + offset
-                            if (checkVector.x !in 0 until width || checkVector.y !in 0 until height) {
+                            if (checkVector.x.toInt() !in 0 until width || checkVector.y.toInt() !in 0 until height) {
                                 return@forEach
                             }
                             val checkColor = colorArray[checkVector]
